server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.136.0",
    "httpx>=0.28.1",
    "lxml>=6.1.0",
    "markdown>=3.10.2",
    "openai>=2.32.0",
    "openai-agents>=0.14.6",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.3.0",
    "python-docx>=1.2.0",
    "sqlalchemy>=2.0.49",
    "weasyprint>=68.1",
]

[dependency-groups]
dev = [
    "loguru>=0.7.3",
    "ruff>=0.15.11",
]
